Remove test scope function calls (Unit, Component, Integration)
Some of them were used incorrectly. Overall not necessary now the
repo has limited scope. The total set of tests is small enough to
be ran in a single run. Therefore, the added complexity is imo not
worth having to explain to potential contributors how to use
testutil to write tests.

If we need this again in the future we can easily add it.