Remove unused GetRepoUser function
The function is currently not used. Typically, the check whether
the user is a member of the repo is done in the server function
for which the user is retrieved. Therefore, there is no usecase
for this function on the client (yet).

If needed, access can be checked with acl Get.