Seal the OrgName so that no invalid org names can exist
Orgnames can only be created inside the package because the
interface contains the private sealed() function. The only
function inside the package that returns orgnames is the
NewOrgName constructor. Because this constructor validates
the name argument it receives we have ensured that any
implementation of the OrgName interface has a validated
orgname.