Add marshal functions to api types
The api types for which primitive arguments are used in communication
with the client, I've added Marshal functions to retrieve the
representation in this primitive type. For api types for which
this type is String, this makes sure we destinguish the function
used for printing: String() and the function used to get the
representation for communication: Marshal.