Refactor generator to Rand with functional options
This introduces more flexibility and reduces the public surface of
the package.

Additionally, I refactored the code to keep the original signature
of the Generator interface and the NewGenerator functions. This to
ensure old code using the package won't break when they update their
dependencies.

Also added a Letters character set and made the io.Reader used as
source of randomness configurable.