Use more future-proof format for credential fingerprints
With this format, we can later use more types without worrying about collisions between the fingerprints of different types. The chance of a sha256 collision is near-zero. However, if for some reason the inputs for 2 fingerprints of different credentials types would be equal, the fingerprint would also become equal.

By introducing this construction for fingerprints, it is impossible for two fingerprints of a different credential type to be equal without having found a SHA-256 collision. The latter would be an awesome, making the resulting server-side error totally worth it.