Remove Fingerprint() from Encryptor and Verifier interface
Getting the fingerprint is now a generalized process for all credential types. Therefore there is no need to implement this function for all credential types separately.

Furthermore, the credential fingerprint is now explicitly passed to any function that creates an account key. The reason for this is that the Encryptor itself has nothing to do with the credential it is tied to. So it makes more sense to decouple this.