Implementation of new EncryptedData spec
As discussed with @mackenbach. This should be forward-compatible to be used for all encrypted data in the API. Not all helper functionality has been implemented, but it is a first start that can be used for account keys.

Comments are still missing, they will be added after a first review of this implementation.