Add godoc for Client.DefaultCredential
Again, this name suffers from the complexity that it isn't the
'default' per se: it also allows configuration from the environment.

It is more the default way to get a credential than that it is the
default credential itself. Small but important difference in meaning.
We'll leave it as is for now and may revisit it later.