Use values instead of pointers for uuid values that are always set
On responses from the server, the UUID values are always given
(except for the parent directory of a dir), therefore, we can return
the values themselves instead of pointers. By doing so, the consuming
applications don't have to dereference the pointers everywhere they
are used.

This change becomes more relevant now, as we are creating endpoints
to get resources by their IDs, so the IDs will be consumed more often
in the future.