Use AuditPaginator instead of generic Paginator
The current paginator is not the right abstraction. There currently
is nothing generic about the paginator, which results in everything
being injected as functions.
Once we implement link headers, we can abstract away how we get the
URL of the next page, which is then done in the same way for all
pages.