Fix session expiration check
The previous check always returned true if the check was performed before the expiration. In most cases this did not hindered the actual functionality of the sessions because the session was just always refreshed. However, if a session was used after it was expired, the client wrongly assumed that the session was still valid, which led to a server error.