Make sure WaitForAuthorizationCode() always returns
No longer mixing different channels, which made it very unclear what happened. By buffering the channels and doing best-effort sending to these channels, we prevent goroutines infinitely hanging on sending on a channel. It is possible that a channel ends up with an unread message after WaitForAuthorizationCode has returned, but that will be cleaned up by the GC.