Refactor access rule encryption
Before this commit, the reencrypter took the first (and only) entry
out of the returned lists. This was possible because the number of
entries in the returned list equals the number of passed accounts
to the function.
With this refactor, the lookup is removed and we can rely fully on
the compiler rather than this implicit agreement that the number of
returned entries equals the number of accounts passed.