Refactor: Explicitly handle err==nil case
Previously, the err==nil case was handled in the
err!=api.ErrNotEncryptedForAccounts block. We prefer the
readability of handling both cases separately, as it's easier
to see this includes the nil case.

This comes at the cost of consiseness and a small performance
penalty of introducing another conditional. We value the
readability more.