Return error for setting a credential
The old behaviour of ignoring the error, resulted in errors during loading a credential (for example, a wrong passphrase), from being ignored. This resulted in getting an "unauthenticated client" error.

Creating an unauthenticated client is no longer needed now that the signup process is removed from the SDK and CLI.